+++ src/structures/src/Constants/maxfloat.c
@@ -1,9 +1,7 @@
 #include <float.h>
 
 
-int maxfloat()
+float maxfloat()
 {
   return( FLT_MAX);
 }
-
-
